+++ Samples~/Scripts/SaintsEditor/GetInSiblingsExample.cs
@@ -1,0 +1,11 @@
+using SaintsField;
+using UnityEngine;
+
+namespace Saintsfield.Samples.Scripts.SaintsEditor
+{
+    public class GetInSiblingsExample : SaintsMonoBehaviour
+    {
+        [GetInSiblings] public SpriteRenderer sr;
+        [GetInSiblings] public SpriteRenderer[] srArray;
+    }
+}